+++ CHANGES.md
@@ -13,6 +13,7 @@ Changes in 11.3
 - Remove null values from URL in PreparePHP_SELF.fnc.php
 - We are inside a sentence, convert nouns to lowercase (except for German) in ListOutput.fnc.php & Prompts.php
 - No "Default for Teacher" checkbox if Type is "Office Only" in AttendanceCodes.php
+- Fix incoherence with AllowEdit() when category_id present or not in URL in Student.php & User.php
 
 Changes in 11.2.3
 -----------------